requirements.txt
streamlit
numpy
pandas
opencv-python-headless
matplotlib
Pillow
pdf2image
PyMuPDF
fpdf
google-generativeai